Don't use -explicit-module-map-file for pcm that has the same name as main module
Currently, dependency scanner returns the command-line to build all
dependency modules including all clang modules. It already contains the
dependency for the clang module that has the same name as the main
module. Don't pass generate explicit module map file for that clang
module. Also the CAS version of the explicit-module-map-file cannot be
used in a `PCM` generation invocation.